Refactor Parameter moved to  Grid namespace, that it can  be reused.
